SingleMolecule: start work on proper subsetting API
This uses the usual FOV[x,y] syntax, but the first argument is a mask
and the second argument is a gene or list of genes. A third argument is
possible and will be used as polygon method: FOV[x, y, "discard"] for
example. FOV[x] will subset just by mask and FOV[y] just by genes.